Внёс необходимые корректировки
