+++ js/main.js
@@ -1,4 +1,3 @@
 import './popup.js';
 import './user-form.js';
-import './form-nonactive.js';
 import './map.js';
