+++ js/user-form.js
@@ -1,7 +1,40 @@
+const adForm = document.querySelector('.ad-form');
+const adFormFields = adForm.querySelectorAll('fieldset');
+const mapFiltersForm = document.querySelector('.map__filters');
+const mapFiltersFormFields = mapFiltersForm.querySelectorAll('.map__filter, .map__features');
 const currentPrice = document.querySelector('#price')
 const houseTypeList = document.querySelector('#type');
 const checkInTime = document.querySelector('#timein');
 const checkOutTime = document.querySelector('#timeout');
+const addressField = adForm.querySelector('#address');
+
+// Form inactive
+
+adForm.classList.add('ad-form--disabled');
+adFormFields.forEach((field) => {
+  field.setAttribute('disabled', '');
+});
+
+mapFiltersForm.classList.add('map__filters--disabled');
+mapFiltersFormFields.forEach((field) => {
+  field.setAttribute('disabled', '');
+});
+
+// Form activation function
+
+const getActivatedForm = () => {
+  adForm.classList.remove('ad-form--disabled');
+  adFormFields.forEach((field) => {
+    field.removeAttribute('disabled', '');
+  });
+  addressField.setAttribute('readonly', '');
+  mapFiltersForm.classList.remove('map__filters--disabled');
+  mapFiltersFormFields.forEach((field) => {
+    field.setAttribute('disabled', '');
+  });
+}
+
+// Check in and check out fields link
 
 const onCheckInTimeChange = () => {
   checkOutTime.value = checkInTime.value;
@@ -10,6 +43,8 @@ const onCheckInTimeChange = () => {
 const onCheckOutTimeChange = () => {
   checkInTime.value = checkOutTime.value;
 }
+
+// house type and price fields link
 
 const onHouseTypeChange = () => {
   currentPrice.placeholder = minPrice[houseTypeList.value];
@@ -34,3 +69,5 @@ houseTypeList.addEventListener('change', onHouseTypeChange);
 checkInTime.value = checkOutTime.value;
 checkInTime.addEventListener('change', onCheckInTimeChange);
 checkOutTime.addEventListener('change', onCheckOutTimeChange);
+
+export {getActivatedForm};
